refactor(frontend): import layout state directly
